Pass estimated song length to `Mixer` class to avoid array padding
Since the destination `AudioSegment`s are no longer precomputed, there's no longer a way for `Mixer` know the track length in advance. As a consequence, we initialize the numpy `ndarray` with zero elements (empty), and then resize it everytime the overlaid segment exceeds the end of the array.

The problem is, resizing an array is a costly operation; there's not a way to do it in-place, so a new array must be created and returned - which takes both time and memory. To avoid this, the song length is now estimated in the `SongRenderer` class, and passed to the `Mixer` so it can pre-allocate an array with that size. It can still resize the array if necessary, but this is much less likely to occur.