Fix wrong track length estimation due to incorrect pitch calculation
The length calculator was being passed notes with the non-weighted pitch values (i.e. F#4 = 45 instead of 0), so the pitch values were ending up much larger. Since we were dividing the sound length by the pitch, it would turn out to be a very small value and the estimated track length would be shorter than the last note's end time.