refactor: replace `pydub.AudioSegment` with our own class
- The `Mixer` class now uses a 2D numpy array for all operations (I don't know why it didn't use one before) instead of a 1D array with one entry per sample *per channel*.
- Sounds are now processed internally as `float32` instead of `int16` (with an oversized `int32` array for mixing), mostly as a consequence of it being the output format for `samplerate`. Fortunately, it is a good outcome - float arrays can just handle the overflow we throw at them.

This commit strives to change as little as possible of the internal workings, as it is a refactor. But this replacement opens up many opportunities for refactoring further aspects of the code, and even making it more efficient - since we now don't rely on handling audio data the way `pydub` expects us to.

Fixes #6