perf: remove `sync` method and apply resampling only once
`nbswave` was 'syncing' all segments to the same sample rate before setting the speed and overlaying them. This is mostly because `pydub` (the library we used before) did it in that way. So we actually resampled most of the notes twice. This, however, is not needed: we can do a single resampling operation by storing the original sound's sample rate and weighing that in when multiplying the speed.

For instance, to make a 48kHz file 1.5x faster with target sample rate 44.1kHz, we do:
`1.5 * 48000 / 44100`